pyproject.toml
[tool.poetry]
name = "hubitatcontrol"
version = "1.1.1"
description = "Hubitat Maker API Interface"
authors = ["Jesse Schoepfer <jelloeater@gmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/Jelloeater/hubitatcontrol"
keywords = ["hubitat", "makerapi","requests"]
classifiers = ["Development Status :: 5 - Production/Stable",
 "Environment :: No Input/Output (Daemon)",
 "Topic :: Home Automation",
 "Intended Audience :: Developers",
 "Operating System :: OS Independent",
]


[tool.poetry.dependencies]
python = "^3.10"
requests = "*"
typer = "*"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = "test/"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "[%(asctime)s] [%(levelname)8s] --- %(message)s (%(filename)s:%(funcName)s():%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.flake8]
exclude =['./tests']
max-line-length = 120
count = false
statistics = true
diff = true
format = "pylint"


[tool.isort]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
line_length = 120
profile = "black"

[tool.tartufo]
repo-path = "."
regex = true
entropy = true
exclude-path-patterns = [
 {path-pattern = 'poetry\.lock'},
 {path-pattern = 'pyproject\.toml'},
 # To not have to escape `\` in regexes, use single quoted
 # TOML 'literal strings'
 {path-pattern = 'docs/source/(.*)\.rst'},
]
exclude-entropy-patterns = [
    {path-pattern = '\.github/workflows/.*\.yml', pattern = 'uses: .*@[a-zA-Z0-9]{40}', reason = 'GitHub Actions'},
    {path-pattern = 'poetry\.lock', pattern = '.'},
    {path-pattern = 'Pipfile\.lock', pattern = '.'},
    {path-pattern = 'README\.md', pattern = '.'},
]

[tool.poetry.group.test.dependencies]
setuptools = "*"
pytest = "*"
pytest-cov = "*"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
pytest-pycharm = "*"
pre-commit = "*"
vulture = "*"
isort = "*"
radon = "*"
xenon = "*"
black = "*"
typing_extensions ="*"
#typeguard ="*" # Typing
#mypy = "*" # Typing
pdoc3 = "*"
pylint = "*"
bandit = "*"
whispers = "*"
mdformat-gfm ="*"
mdformat-frontmatter ="*"
flake8 = "*"
Flake8-pyproject="*"
tartufo="*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
